refactor: move ResourcePageLayout
