+++ src/components/TeacherComponents/ResourcePageLayout/index.ts
